# scrumpoker_api/internal/domain/history/history_gen.go
// Package history provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.9.1 DO NOT EDIT.
package history

import (
	"fmt"
	"net/http"

	externalRef0 "github.com/chack93/scrumpoker_api/internal/domain/common"
	"github.com/deepmap/oapi-codegen/pkg/runtime"
	"github.com/labstack/echo/v4"
)

// History defines model for History.
type History struct {
	// Embedded struct due to allOf(../common/common.yaml#/components/schemas/BaseModel)
	externalRef0.BaseModel `yaml:",inline"`
	// Embedded struct due to allOf(#/components/schemas/HistoryNew)
	HistoryNew `yaml:",inline"`
	// Embedded fields due to inline allOf schema
	GameId *string `json:"gameId,omitempty"`
}

// HistoryList defines model for HistoryList.
type HistoryList []History

// HistoryNew defines model for HistoryNew.
type HistoryNew struct {
	// Embedded fields due to inline allOf schema
	ClientId   *string `json:"clientId,omitempty"`
	ClientName *string `json:"clientName,omitempty"`
	Estimation *string `json:"estimation,omitempty"`
	SessionId  *string `json:"sessionId,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /history/{sessionId})
	ListHistory(ctx echo.Context, sessionId string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// ListHistory converts echo context to params.
func (w *ServerInterfaceWrapper) ListHistory(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "sessionId" -------------
	var sessionId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "sessionId", runtime.ParamLocationPath, ctx.Param("sessionId"), &sessionId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter sessionId: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.ListHistory(ctx, sessionId)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/history/:sessionId", wrapper.ListHistory)

}